Implement new game and draw features in Solitaire
The main functions added in this commit include a new game and draw feature in the Solitaire code. The new game function empties the deck and creates a new, shuffled deck. The draw function has been modified to allow reusing of the discarded cards once the deck is depleted. There's also a small refactoring of the UI code for better layout and coherency.