Refactor card detail and add new game functions in Solitaire
This commit involves refactoring of Card Detail in Solitaire app to offer a more detailed view. CardDetail function was created, which can be turned on and off by passing 'showFullDetail' boolean. Additionally, the commit adds new game features such as dialog for game confirmation as well as pause and resume timer functions for the game. Score implementation is enhanced, it updates based on card moves while keeping track of time.