Fix issue with card removal logic
The commit rectifies an error in the method responsible for the removal of cards in the list. Previously, the code mistakenly dropped the last 'num' cards from the list instead of dropping 'num' number of cards from wherever they are in the list. This bug was causing an incorrect score calculation when re-adding cards to the deck.